Enum config_helloworld::Error

#[non_exhaustive]
pub(crate) enum Error {

    ConformancePackTemplateValidationException(ConformancePackTemplateValidationException),
    InsufficientDeliveryPolicyException(InsufficientDeliveryPolicyException),
    InsufficientPermissionsException(InsufficientPermissionsException),
    InvalidConfigurationRecorderNameException(InvalidConfigurationRecorderNameException),
    InvalidDeliveryChannelNameException(InvalidDeliveryChannelNameException),
    InvalidExpressionException(InvalidExpressionException),
    InvalidLimitException(InvalidLimitException),
    InvalidNextTokenException(InvalidNextTokenException),
    InvalidParameterValueException(InvalidParameterValueException),
    InvalidRecordingGroupException(InvalidRecordingGroupException),
    InvalidResultTokenException(InvalidResultTokenException),
    InvalidRoleException(InvalidRoleException),
    InvalidS3KeyPrefixException(InvalidS3KeyPrefixException),
    InvalidS3KmsKeyArnException(InvalidS3KmsKeyArnException),
    InvalidSnsTopicArnException(InvalidSnsTopicArnException),
    InvalidTimeRangeException(InvalidTimeRangeException),
    LastDeliveryChannelDeleteFailedException(LastDeliveryChannelDeleteFailedException),
    LimitExceededException(LimitExceededException),
    MaxActiveResourcesExceededException(MaxActiveResourcesExceededException),
    MaxNumberOfConfigRulesExceededException(MaxNumberOfConfigRulesExceededException),
    MaxNumberOfConfigurationRecordersExceededException(MaxNumberOfConfigurationRecordersExceededException),
    MaxNumberOfConformancePacksExceededException(MaxNumberOfConformancePacksExceededException),
    MaxNumberOfDeliveryChannelsExceededException(MaxNumberOfDeliveryChannelsExceededException),
    MaxNumberOfOrganizationConfigRulesExceededException(MaxNumberOfOrganizationConfigRulesExceededException),
    MaxNumberOfOrganizationConformancePacksExceededException(MaxNumberOfOrganizationConformancePacksExceededException),
    MaxNumberOfRetentionConfigurationsExceededException(MaxNumberOfRetentionConfigurationsExceededException),
    NoAvailableConfigurationRecorderException(NoAvailableConfigurationRecorderException),
    NoAvailableDeliveryChannelException(NoAvailableDeliveryChannelException),
    NoAvailableOrganizationException(NoAvailableOrganizationException),
    NoRunningConfigurationRecorderException(NoRunningConfigurationRecorderException),
    NoSuchBucketException(NoSuchBucketException),
    NoSuchConfigRuleException(NoSuchConfigRuleException),
    NoSuchConfigRuleInConformancePackException(NoSuchConfigRuleInConformancePackException),
    NoSuchConfigurationAggregatorException(NoSuchConfigurationAggregatorException),
    NoSuchConfigurationRecorderException(NoSuchConfigurationRecorderException),
    NoSuchConformancePackException(NoSuchConformancePackException),
    NoSuchDeliveryChannelException(NoSuchDeliveryChannelException),
    NoSuchOrganizationConfigRuleException(NoSuchOrganizationConfigRuleException),
    NoSuchOrganizationConformancePackException(NoSuchOrganizationConformancePackException),
    NoSuchRemediationConfigurationException(NoSuchRemediationConfigurationException),
    NoSuchRemediationExceptionException(NoSuchRemediationExceptionException),
    NoSuchRetentionConfigurationException(NoSuchRetentionConfigurationException),
    OrganizationAccessDeniedException(OrganizationAccessDeniedException),
    OrganizationAllFeaturesNotEnabledException(OrganizationAllFeaturesNotEnabledException),
    OrganizationConformancePackTemplateValidationException(OrganizationConformancePackTemplateValidationException),
    OversizedConfigurationItemException(OversizedConfigurationItemException),
    RemediationInProgressException(RemediationInProgressException),
    ResourceConcurrentModificationException(ResourceConcurrentModificationException),
    ResourceInUseException(ResourceInUseException),
    ResourceNotDiscoveredException(ResourceNotDiscoveredException),
    ResourceNotFoundException(ResourceNotFoundException),
    TooManyTagsException(TooManyTagsException),
    ValidationException(ValidationException),
    Unhandled(Box<dyn Error + Send + Sync + 'static, Global>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ConformancePackTemplateValidationException(ConformancePackTemplateValidationException)

You have specified a template that is not valid or supported.

Tuple Fields of ConformancePackTemplateValidationException

0: ConformancePackTemplateValidationException

InsufficientDeliveryPolicyException(InsufficientDeliveryPolicyException)

Your Amazon S3 bucket policy does not permit Config to write to it.

Tuple Fields of InsufficientDeliveryPolicyException

0: InsufficientDeliveryPolicyException

InsufficientPermissionsException(InsufficientPermissionsException)

Indicates one of the following errors:

For PutConfigRule, the rule cannot be created because the IAM role assigned to Config lacks permissions to perform the config:Put* action.

For PutConfigRule, the Lambda function cannot be invoked. Check the function ARN, and check the function's permissions.

For PutOrganizationConfigRule, organization config rule cannot be created because you do not have permissions to call IAM GetRole action or create a service linked role.

For PutConformancePack and PutOrganizationConformancePack, a conformance pack cannot be created because you do not have permissions:

To call IAM GetRole action or create a service linked role.

To read Amazon S3 bucket.

Tuple Fields of InsufficientPermissionsException

0: InsufficientPermissionsException

InvalidConfigurationRecorderNameException(InvalidConfigurationRecorderNameException)

You have provided a configuration recorder name that is not valid.

Tuple Fields of InvalidConfigurationRecorderNameException

0: InvalidConfigurationRecorderNameException

InvalidDeliveryChannelNameException(InvalidDeliveryChannelNameException)

The specified delivery channel name is not valid.

Tuple Fields of InvalidDeliveryChannelNameException

0: InvalidDeliveryChannelNameException

InvalidExpressionException(InvalidExpressionException)

The syntax of the query is incorrect.

Tuple Fields of InvalidExpressionException

0: InvalidExpressionException

InvalidLimitException(InvalidLimitException)

The specified limit is outside the allowable range.

Tuple Fields of InvalidLimitException

0: InvalidLimitException

InvalidNextTokenException(InvalidNextTokenException)

The specified next token is invalid. Specify the nextToken string that was returned in the previous response to get the next page of results.

Tuple Fields of InvalidNextTokenException

0: InvalidNextTokenException

InvalidParameterValueException(InvalidParameterValueException)

One or more of the specified parameters are invalid. Verify that your parameters are valid and try again.

Tuple Fields of InvalidParameterValueException

0: InvalidParameterValueException

InvalidRecordingGroupException(InvalidRecordingGroupException)

Config throws an exception if the recording group does not contain a valid list of resource types. Invalid values might also be incorrectly formatted.

Tuple Fields of InvalidRecordingGroupException

0: InvalidRecordingGroupException

InvalidResultTokenException(InvalidResultTokenException)

The specified ResultToken is invalid.

Tuple Fields of InvalidResultTokenException

0: InvalidResultTokenException

InvalidRoleException(InvalidRoleException)

You have provided a null or empty role ARN.

Tuple Fields of InvalidRoleException

0: InvalidRoleException

InvalidS3KeyPrefixException(InvalidS3KeyPrefixException)

The specified Amazon S3 key prefix is not valid.

Tuple Fields of InvalidS3KeyPrefixException

0: InvalidS3KeyPrefixException

InvalidS3KmsKeyArnException(InvalidS3KmsKeyArnException)

The specified Amazon KMS Key ARN is not valid.

Tuple Fields of InvalidS3KmsKeyArnException

0: InvalidS3KmsKeyArnException

InvalidSnsTopicArnException(InvalidSnsTopicArnException)

The specified Amazon SNS topic does not exist.

Tuple Fields of InvalidSnsTopicArnException

0: InvalidSnsTopicArnException

InvalidTimeRangeException(InvalidTimeRangeException)

The specified time range is not valid. The earlier time is not chronologically before the later time.

Tuple Fields of InvalidTimeRangeException

0: InvalidTimeRangeException

LastDeliveryChannelDeleteFailedException(LastDeliveryChannelDeleteFailedException)

You cannot delete the delivery channel you specified because the configuration recorder is running.

Tuple Fields of LastDeliveryChannelDeleteFailedException

0: LastDeliveryChannelDeleteFailedException

LimitExceededException(LimitExceededException)

For StartConfigRulesEvaluation API, this exception is thrown if an evaluation is in progress or if you call the StartConfigRulesEvaluation API more than once per minute.

For PutConfigurationAggregator API, this exception is thrown if the number of accounts and aggregators exceeds the limit.

Tuple Fields of LimitExceededException

0: LimitExceededException

MaxActiveResourcesExceededException(MaxActiveResourcesExceededException)

You have reached the limit (100,000) of active custom resource types in your account. Delete unused resources using DeleteResourceConfig.

Tuple Fields of MaxActiveResourcesExceededException

0: MaxActiveResourcesExceededException

MaxNumberOfConfigRulesExceededException(MaxNumberOfConfigRulesExceededException)

Failed to add the Config rule because the account already contains the maximum number of 150 rules. Consider deleting any deactivated rules before you add new rules.

Tuple Fields of MaxNumberOfConfigRulesExceededException

0: MaxNumberOfConfigRulesExceededException

MaxNumberOfConfigurationRecordersExceededException(MaxNumberOfConfigurationRecordersExceededException)

You have reached the limit of the number of recorders you can create.

Tuple Fields of MaxNumberOfConfigurationRecordersExceededException

0: MaxNumberOfConfigurationRecordersExceededException

MaxNumberOfConformancePacksExceededException(MaxNumberOfConformancePacksExceededException)

You have reached the limit (6) of the number of conformance packs in an account (6 conformance pack with 25 Config rules per pack).

Tuple Fields of MaxNumberOfConformancePacksExceededException

0: MaxNumberOfConformancePacksExceededException

MaxNumberOfDeliveryChannelsExceededException(MaxNumberOfDeliveryChannelsExceededException)

You have reached the limit of the number of delivery channels you can create.

Tuple Fields of MaxNumberOfDeliveryChannelsExceededException

0: MaxNumberOfDeliveryChannelsExceededException

MaxNumberOfOrganizationConfigRulesExceededException(MaxNumberOfOrganizationConfigRulesExceededException)

You have reached the limit of the number of organization config rules you can create.

Tuple Fields of MaxNumberOfOrganizationConfigRulesExceededException

0: MaxNumberOfOrganizationConfigRulesExceededException

MaxNumberOfOrganizationConformancePacksExceededException(MaxNumberOfOrganizationConformancePacksExceededException)

You have reached the limit (6) of the number of organization conformance packs in an account (6 conformance pack with 25 Config rules per pack per account).

Tuple Fields of MaxNumberOfOrganizationConformancePacksExceededException

0: MaxNumberOfOrganizationConformancePacksExceededException

MaxNumberOfRetentionConfigurationsExceededException(MaxNumberOfRetentionConfigurationsExceededException)

Failed to add the retention configuration because a retention configuration with that name already exists.

Tuple Fields of MaxNumberOfRetentionConfigurationsExceededException

0: MaxNumberOfRetentionConfigurationsExceededException

NoAvailableConfigurationRecorderException(NoAvailableConfigurationRecorderException)

There are no configuration recorders available to provide the role needed to describe your resources. Create a configuration recorder.

Tuple Fields of NoAvailableConfigurationRecorderException

0: NoAvailableConfigurationRecorderException

NoAvailableDeliveryChannelException(NoAvailableDeliveryChannelException)

There is no delivery channel available to record configurations.

Tuple Fields of NoAvailableDeliveryChannelException

0: NoAvailableDeliveryChannelException

NoAvailableOrganizationException(NoAvailableOrganizationException)

Organization is no longer available.

Tuple Fields of NoAvailableOrganizationException

0: NoAvailableOrganizationException

NoRunningConfigurationRecorderException(NoRunningConfigurationRecorderException)

There is no configuration recorder running.

Tuple Fields of NoRunningConfigurationRecorderException

0: NoRunningConfigurationRecorderException

NoSuchBucketException(NoSuchBucketException)

The specified Amazon S3 bucket does not exist.

Tuple Fields of NoSuchBucketException

0: NoSuchBucketException

NoSuchConfigRuleException(NoSuchConfigRuleException)

One or more Config rules in the request are invalid. Verify that the rule names are correct and try again.

Tuple Fields of NoSuchConfigRuleException

0: NoSuchConfigRuleException

NoSuchConfigRuleInConformancePackException(NoSuchConfigRuleInConformancePackException)

Config rule that you passed in the filter does not exist.

Tuple Fields of NoSuchConfigRuleInConformancePackException

0: NoSuchConfigRuleInConformancePackException

NoSuchConfigurationAggregatorException(NoSuchConfigurationAggregatorException)

You have specified a configuration aggregator that does not exist.

Tuple Fields of NoSuchConfigurationAggregatorException

0: NoSuchConfigurationAggregatorException

NoSuchConfigurationRecorderException(NoSuchConfigurationRecorderException)

You have specified a configuration recorder that does not exist.

Tuple Fields of NoSuchConfigurationRecorderException

0: NoSuchConfigurationRecorderException

NoSuchConformancePackException(NoSuchConformancePackException)

You specified one or more conformance packs that do not exist.

Tuple Fields of NoSuchConformancePackException

0: NoSuchConformancePackException

NoSuchDeliveryChannelException(NoSuchDeliveryChannelException)

You have specified a delivery channel that does not exist.

Tuple Fields of NoSuchDeliveryChannelException

0: NoSuchDeliveryChannelException

NoSuchOrganizationConfigRuleException(NoSuchOrganizationConfigRuleException)

You specified one or more organization config rules that do not exist.

Tuple Fields of NoSuchOrganizationConfigRuleException

0: NoSuchOrganizationConfigRuleException

NoSuchOrganizationConformancePackException(NoSuchOrganizationConformancePackException)

Config organization conformance pack that you passed in the filter does not exist.

For DeleteOrganizationConformancePack, you tried to delete an organization conformance pack that does not exist.

Tuple Fields of NoSuchOrganizationConformancePackException

0: NoSuchOrganizationConformancePackException

NoSuchRemediationConfigurationException(NoSuchRemediationConfigurationException)

You specified an Config rule without a remediation configuration.

Tuple Fields of NoSuchRemediationConfigurationException

0: NoSuchRemediationConfigurationException

NoSuchRemediationExceptionException(NoSuchRemediationExceptionException)

You tried to delete a remediation exception that does not exist.

Tuple Fields of NoSuchRemediationExceptionException

0: NoSuchRemediationExceptionException

NoSuchRetentionConfigurationException(NoSuchRetentionConfigurationException)

You have specified a retention configuration that does not exist.

Tuple Fields of NoSuchRetentionConfigurationException

0: NoSuchRetentionConfigurationException

OrganizationAccessDeniedException(OrganizationAccessDeniedException)

For PutConfigurationAggregator API, you can see this exception for the following reasons:

No permission to call EnableAWSServiceAccess API

The configuration aggregator cannot be updated because your Amazon Web Services Organization management account or the delegated administrator role changed. Delete this aggregator and create a new one with the current Amazon Web Services Organization.

The configuration aggregator is associated with a previous Amazon Web Services Organization and Config cannot aggregate data with current Amazon Web Services Organization. Delete this aggregator and create a new one with the current Amazon Web Services Organization.

You are not a registered delegated administrator for Config with permissions to call ListDelegatedAdministrators API. Ensure that the management account registers delagated administrator for Config service principle name before the delegated administrator creates an aggregator.

For all OrganizationConfigRule and OrganizationConformancePack APIs, Config throws an exception if APIs are called from member accounts. All APIs must be called from organization master account.

Tuple Fields of OrganizationAccessDeniedException

0: OrganizationAccessDeniedException

OrganizationAllFeaturesNotEnabledException(OrganizationAllFeaturesNotEnabledException)

Config resource cannot be created because your organization does not have all features enabled.

Tuple Fields of OrganizationAllFeaturesNotEnabledException

0: OrganizationAllFeaturesNotEnabledException

OrganizationConformancePackTemplateValidationException(OrganizationConformancePackTemplateValidationException)

You have specified a template that is not valid or supported.

Tuple Fields of OrganizationConformancePackTemplateValidationException

0: OrganizationConformancePackTemplateValidationException

OversizedConfigurationItemException(OversizedConfigurationItemException)

The configuration item size is outside the allowable range.

Tuple Fields of OversizedConfigurationItemException

0: OversizedConfigurationItemException

RemediationInProgressException(RemediationInProgressException)

Remediation action is in progress. You can either cancel execution in Amazon Web Services Systems Manager or wait and try again later.

Tuple Fields of RemediationInProgressException

0: RemediationInProgressException

ResourceConcurrentModificationException(ResourceConcurrentModificationException)

Two users are trying to modify the same query at the same time. Wait for a moment and try again.

Tuple Fields of ResourceConcurrentModificationException

0: ResourceConcurrentModificationException

ResourceInUseException(ResourceInUseException)

You see this exception in the following cases:

For DeleteConfigRule, Config is deleting this rule. Try your request again later.

For DeleteConfigRule, the rule is deleting your evaluation results. Try your request again later.

For DeleteConfigRule, a remediation action is associated with the rule and Config cannot delete this rule. Delete the remediation action associated with the rule before deleting the rule and try your request again later.

For PutConfigOrganizationRule, organization config rule deletion is in progress. Try your request again later.

For DeleteOrganizationConfigRule, organization config rule creation is in progress. Try your request again later.

For PutConformancePack and PutOrganizationConformancePack, a conformance pack creation, update, and deletion is in progress. Try your request again later.

For DeleteConformancePack, a conformance pack creation, update, and deletion is in progress. Try your request again later.

Tuple Fields of ResourceInUseException

0: ResourceInUseException

ResourceNotDiscoveredException(ResourceNotDiscoveredException)

You have specified a resource that is either unknown or has not been discovered.

Tuple Fields of ResourceNotDiscoveredException

0: ResourceNotDiscoveredException

ResourceNotFoundException(ResourceNotFoundException)

You have specified a resource that does not exist.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

TooManyTagsException(TooManyTagsException)

You have reached the limit of the number of tags you can use. You have more than 50 tags.

Tuple Fields of TooManyTagsException

0: TooManyTagsException

ValidationException(ValidationException)

The requested action is not valid.

For PutStoredQuery, you will see this exception if there are missing required fields or if the input value fails the validation, or if you are trying to create more than 300 queries.

For GetStoredQuery, ListStoredQuery, and DeleteStoredQuery you will see this exception if there are missing required fields or if the input value fails the validation.

Tuple Fields of ValidationException

0: ValidationException

Unhandled(Box<dyn Error + Send + Sync + 'static, Global>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static, Global>